Improve document subcmd file naming
